requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain_anthropic
python-dotenv
langchainhub
pymongo
yfinance